feat(dependencies): add rollup and related plugins for build setup
Added Rollup and associated plugins such as rollup-plugin-replace and
rollup-plugin-postcss to the project's devDependencies. This indicates that
the project is now utilizing Rollup for its build configuration, potentially
for bundling and optimizing assets.

Additionally, introduced less and @web/dev-server-rollup to support Less CSS
preprocessing and improved the test command to include Puppeteer.